Remove useless OnChainWalletSetupService
