+++ BTCPayServer/Plugins/Wallets/WalletsPlugin.cs
@@ -1,6 +1,5 @@
 using BTCPayServer.Abstractions.Models;
 using BTCPayServer.Client;
-using BTCPayServer.Plugins.Wallets.Services;
 using BTCPayServer.Services;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.DependencyInjection.Extensions;
@@ -18,7 +17,6 @@ public class WalletsPlugin : BaseBTCPayServerPlugin
     public override void Execute(IServiceCollection services)
     {
         services.AddSearchResultItemProvider<WalletSearchResultProvider>();
-        services.TryAddSingleton<OnChainWalletSetupService>();
         services.AddTransient<HotwalletSafe>();
         services.AddPolicyDefinitions(
             new PolicyDefinition(
